Use threads and enable ICMP correction by default
Partial re-write to use two threads - one sending pings and the other receiving. This enabled the timeout to be set longer than the ping interval.

This patch also enables the correction of ICMP timestamps by default. Raw timestamps can now be used by specifying the -r option, and the option to use Unix timestamps in the log is now -u.